Mostly finished with DeleteBook. Changes to TransactionFactory As always, cleaning to be done.
A few dirty things are going on in the ListBookView and the DeleteBooksTransaction and ListBooksTransaction now in order for the view to dynamically change in relation to whether or not the user is trying to modify or delete a book.
This might be unavoidable, or it might not. For now, things work.

This commit also includes changes the Transactions are created (where they are created) and what they can be passed as parameters.